% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArcPlot.R
\name{ArcPlot}
\alias{ArcPlot}
\title{Ploting or adding the arc connected two points}
\usage{
ArcPlot(
  p1,
  p2,
  convex = TRUE,
  return = FALSE,
  angle = 90,
  add = TRUE,
  radian = NA,
  col = "black",
  lty = 1,
  lwd = 1
)
}
\arguments{
\item{p1}{a coordinate (x,y) vectors of one point.}

\item{p2}{a coordinate (x,y) vectors of the other point.}

\item{convex}{logical value indicating the convex arc (up/right) or concave arc(down/left).}

\item{return}{logical value indicating the list was returned or not.}

\item{angle}{numeric (<90) indicating the degree of the central Angle corresponding to the arc,.}

\item{add}{logical value indicating adding this arc to currents plot or not.}

\item{radian}{numeric indicating the radian of arc you need to plot, default is NA.
It will cover the value of angle.(if you do not want to supply the angle of arc).}

\item{...}{Further graphical parameters may also be supplied as arguments, including, line type, lty, line width, lwd, color, col.}
}
\value{
Adding the arc to current plot or plot the arc in new device.
A list named 'curve_expr' including: information of this arc,
cc_current(vector of numeric), the coordinates of center of circle,
radius(numeric), the radius of circle
exprs(vector of characters), the expression of piecewise function of this arc
xrange(dataframe), the x range of the corresponding piecewise function (per row means the range).
}
\description{
A generic function ploting or adding the arc linking two points.
}
\examples{
#example 1
plot(0,0,col="white",bty="o",xlim=c(-10,10),ylim=c(-10,10))
p1 <- c(-7,5)
p2 <- c(-3,5)
points(c(p1[1],p2[1]),c(p1[2],p2[2]))
ArcPlot(p1,p2,convex = TRUE,angle = 90)
m <- ArcPlot(p1,p2,convex = FALSE,return=TRUE,angle = 200)
points(-5,CalArcY(-5,m),col="red")

#example 2
#cat data
cat_da <- data.frame(c(0,0,-1,1,-2,-1,1,2,-2,-3,-1.5,1.5,3,2,-3,0,3,0),
  c(0.5,-0.5,-0.5,-0.5,1.5,1.5,1.5,1.5,2.5,4,3,3,4,2.5,0,3.6,0,-2.5))
plot(0,0,bty="o",xlim=c(-5,5),ylim=c(-5,5),col="white")
#nose
ArcPlot(c(cat_da[1,1],cat_da[1,2]),c(cat_da[2,1],cat_da[2,2]),convex = TRUE,angle = 360)
##mouse
ArcPlot(c(cat_da[2,1],cat_da[2,2]),c(cat_da[3,1],cat_da[3,2]),convex = FALSE)
ArcPlot(c(cat_da[2,1],cat_da[2,2]),c(cat_da[4,1],cat_da[4,2]),convex = FALSE)
#eye
ArcPlot(c(cat_da[5,1],cat_da[5,2]),c(cat_da[6,1],cat_da[6,2]),convex = TRUE)
ArcPlot(c(cat_da[7,1],cat_da[7,2]),c(cat_da[8,1],cat_da[8,2]),convex = TRUE)
#outline
left_face <- ArcPlot(c(cat_da[15,1],cat_da[15,2]),c(cat_da[16,1],cat_da[16,2]),convex = TRUE,return=TRUE)
right_face <- ArcPlot(c(cat_da[16,1],cat_da[16,2]),c(cat_da[17,1],cat_da[17,2]),convex = TRUE,return=TRUE)
#left ear
ArcPlot(c(cat_da[9,1],CalArcY(cat_da[9,1],left_face)),c(cat_da[10,1],cat_da[10,2]),convex = FALSE)
ArcPlot(c(cat_da[9,1]-0.5,CalArcY(cat_da[9,1]-0.5,left_face)),c(cat_da[10,1],cat_da[10,2]),convex = FALSE)
ArcPlot(c(cat_da[10,1],cat_da[10,2]),c(cat_da[11,1],CalArcY(cat_da[11,1],left_face)),convex = TRUE)
#right ear
ArcPlot(c(cat_da[12,1],CalArcY(cat_da[12,1],right_face)),c(cat_da[13,1],cat_da[13,2]),convex = TRUE)
ArcPlot(c(cat_da[13,1],cat_da[13,2]),c(cat_da[14,1],CalArcY(cat_da[14,1],right_face)),convex = FALSE)
ArcPlot(c(cat_da[13,1],cat_da[13,2]),c(cat_da[14,1]+0.5,CalArcY(cat_da[14,1]+.5,right_face)),convex = FALSE)
#outline
ArcPlot(c(cat_da[15,1],cat_da[15,2]),c(cat_da[18,1],cat_da[18,2]),convex = FALSE)
ArcPlot(c(cat_da[17,1],cat_da[17,2]),c(cat_da[18,1],cat_da[18,2]),convex = FALSE)
#hair
lines(c(0,0),c(CalArcY(0,left_face),2.5))
lines(c(-0.5,-0.5),c(CalArcY(-.5,left_face),2.5))
lines(c(0.5,0.5),c(CalArcY(.5,right_face),2.5))
#beard
lines(c(-1.5,-4),c(0,0.6))
lines(c(1.5,4),c(0,0.6))
lines(c(-1.5,-4),c(-0.4,-0.6))
lines(c(1.5,4),c(-0.4,-0.6))
lines(c(-1.5,-4),c(-0.8,-1.8))
lines(c(1.5,4),c(-0.8,-1.8))

}
\author{
Yi Li  \email{gugeliyiyi@gmail.com; https://github.com/liyi522/ArcPlot/}
}
