% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalArcY.R
\name{CalArcY}
\alias{CalArcY}
\title{Calculate the corresponding y value of arc based on x which input}
\usage{
CalArcY(x = NA, arcplot = list())
}
\arguments{
\item{x}{numeric, the x value of the point you want to get.}

\item{arcplot}{list, returned by the ArcPlot function.}
}
\value{
The vector of corresponding y values (1 or 2 numeric variables).
A list named 'curve_expr' including: information of this arc,
}
\description{
A function to calculate the corresponding y value for the output of ArcPlot function based on the x input.
}
\examples{
plot(0,0,col="white",bty="o",xlim=c(-10,10),ylim=c(-10,10))
p1 <- c(-7,5)
p2 <- c(-3,5)
points(c(p1[1],p2[1]),c(p1[2],p2[2]))
ArcPlot(p1,p2,convex = TRUE,return=FALSE,angle = 90)
m <- ArcPlot(p1,p2,convex = FALSE,angle = 200)
points(-5,CalArcY(-5,m),col="red")
}
\author{
Yi Li  \email{gugeliyiyi@gmail.com; https://github.com/liyi522/ArcPlot/}
}
